add multibyte public version support (zcash support) (#9)
* add multibyte public version support
- adds support for coins, such as zcash, that use two or more bytes for
versioning of public addresses. This was done by converting all versions
passed as Buffers to coinstring, which the library insists on for
multibyte support.
- added zcash into test cases, required adding a conditional whereby
versions are passed into `fromWif` in the case the coin doesn't follow
the previous convention of priv - pub = 0x80. Can be easily done by also
leveraging coininfo.

* add support for pub version with leading zero byte(s)
- fixes previously incorrect length formula in `beUIntToBuffer`. Was
off-by-one for %256.
- stricter type checking on bufferizeVersion. Can handle BE uint, or in
case of LE uint or hex bytes with leading zeroes, a direct hex string is
expected to be provided.
- added fake multibytetestcoin, sharing same private byte as bitcoin,
but containing public version with a leading 00 byte. Public prefix
should result in `1337D` (chosen by coincidence using bs58check :/, but
since first char would be 1 due to 00, it seemed to look
nicest/different/obvious).

* improves new utils codebase
- implementing @fanatid recommended changes
- streamlining new utils function codebase, changing hex sequence
parsing from regexp to slicing.
- added non 0x prefixed multibytetestcoin version into fixtures.